# Enforce type stability for cmaes