--- a/src/cmaes.jl
+++ b/src/cmaes.jl
@@ -33,7 +33,8 @@ mutable struct CMAESState{T, TI, TT} <: AbstractOptimizerState
     fittest::TI
     function CMAESState(N::Int, τ::T1, τ_c::T2, τ_σ::T3, fitpop::Vector{T}, C::Matrix{T},
         s::Vector{T}, s_σ::Vector{T}, σ::T, parent::TI, fittest::TI) where {T, TI, T1, T2, T3}
-        new{T,TI,promote_type(T1,T2,T3)}(N, promote(τ, τ_c, τ_σ)..., fitpop, C,
+        T = promote_type(T1,T2,T3)
+        new{T,TI,T}(N, T(τ), T(τ_c), T(τ_σ), fitpop, C,
             s, s_σ, σ, parent, fittest)
     end
 end