The interfaces after refactoring are very nice!
The performance can be improved a lot if type stability are enforced.

## Test Code
```julia
using Random
Random.seed!(2)
N = 2
rosenbrock(x) = (1.0 - x[1])^2 + 100.0 * (x[2] - x[1]^2)^2
using Evolutionary
# termination condition
using BenchmarkTools
@benchmark result = Evolutionary.optimize(rosenbrock, randn(2), $(CMAES(;μ=5, λ=100)))
```

## Before
```julia repl
BenchmarkTools.Trial: 
  memory estimate:  34.43 MiB
  allocs estimate:  1124405
  --------------
  minimum time:     132.045 ms (4.46% GC)
  median time:      420.937 ms (4.47% GC)
  mean time:        444.512 ms (3.98% GC)
  maximum time:     757.132 ms (3.90% GC)
  --------------
  samples:          12
  evals/sample:     1
```

## After
```julia repl
BenchmarkTools.Trial: 
  memory estimate:  17.97 MiB
  allocs estimate:  207935
  --------------
  minimum time:     26.536 ms (0.00% GC)
  median time:      67.595 ms (11.88% GC)
  mean time:        66.158 ms (9.74% GC)
  maximum time:     96.395 ms (13.28% GC)
  --------------
  samples:          76
  evals/sample:     1
```