@GiggleLiu 2020-05-03
NOTE: I haven't checked other methods. Maybe other methods can be improved in a similar approach.